+++ docs/changelog.md
@@ -43,6 +43,10 @@
 
 ## 更新说明
 
+1.8.25 fix: fix stock_circulate_stock_holder interface
+
+    1. 修复 stock_circulate_stock_holder 接口
+
 1.8.24 fix: fix stock_ipo_benefit_ths interface
 
     1. 修复 stock_ipo_benefit_ths 接口
@@ -1630,6 +1634,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.8.25 fix: fix stock_circulate_stock_holder interface
 
 1.8.24 fix: fix stock_ipo_benefit_ths interface